refactor(engine): extract shared PTY and terminal state crates
