+++ diri/crates/diri-engine/Cargo.toml
@@ -9,14 +9,37 @@ license.workspace = true
 alacritty_terminal.workspace = true
 base64.workspace = true
 diri-proto.workspace = true
+diri-pty.workspace = true
+diri-terminal-state.workspace = true
 getrandom.workspace = true
 plist.workspace = true
 regex.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+sha2.workspace = true
 
 [target.'cfg(unix)'.dependencies]
 libc = "0.2"
+
+[target.'cfg(target_os = "macos")'.dependencies]
+objc2 = "=0.6.4"
+objc2-app-kit = { version = "=0.3.2", default-features = false, features = [
+    "NSAlert",
+    "NSApplication",
+    "NSButton",
+    "NSCell",
+    "NSControl",
+    "NSResponder",
+    "NSRunningApplication",
+    "NSSecureTextField",
+    "NSTextField",
+    "NSView",
+] }
+objc2-foundation = { version = "=0.3.2", default-features = false, features = [
+    "NSGeometry",
+    "NSObject",
+    "NSString",
+] }
 
 [dev-dependencies]
 tempfile.workspace = true
