+++ diri/crates/diri-pty/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "diri-pty"
+version = "0.1.0"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+
+[target.'cfg(unix)'.dependencies]
+libc = "0.2"
